POC - add embassy based runtime
This commits adds an embassy based rust runtime inside a Async Runnable.
Currently this is limited to only one Runtime, but the concept is
extendable to host one runtime for each FreeRtos task / priority level.
In the near future a forked runtime would be interesting which returns
control at every await point to ensure the cpp runnables do not get
starved by the embassy runnable.

This is currently only a POC. Code quality is lacking. The S32K148 binary
is not working. Do not merge!

Change-Id: I38e1e1d0d2298e2bc77c700d5d04c920fa42c74d